init.mako.rc: update cpufreq permissions for BCM
Since the permissions for interactive are in init.rc we need to set our custom permissions
when the boot has completed since we are sure our permissions will stick. This is needed as
the touchboost implimentation used in Interactive writes to user space files and needs the
necessary permissions changes.

Simple stuff like setting the default govenor to interactive after fixing it is an obious choice

Also set the lowest cpufreq to 162Mhz as thats the lowest in BCM

Patchset 2: Fix derp

Change-Id: I59b3fd2cf6a60bc70dc210ef08e28dee107454db